+++ src/App/Documentation/components/Topbar/index.js
@@ -1,9 +1,8 @@
 import React, { Component } from "react";
 
-import ComponentPreview from "../../utils/ComponentPreview";
-import DocToc from "../../utils/DocToc";
-import TopbarComponent from "../../../components/Topbar";
-import { topbar } from "../../../../px-script";
+import { ComponentPreview, DocToc } from "#";
+import TopbarComponent from "@/Topbar";
+import { topbar } from "$/px-script";
 
 const menu = {
     btn: {