+++ src/App/Documentation/utilities/Display/index.js
@@ -1,8 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import ComponentPreview from "../../utils/ComponentPreview";
-import DocToc from "../../utils/DocToc";
+import { ComponentPreview, DocToc } from "#";
 
 const HowItWorks = () => (
     <>