+++ src/App/Documentation/utils/index.js
@@ -1,0 +1,13 @@
+import ComponentPreview, { Attribute, PxScript } from "./ComponentPreview";
+import DocHeading from "./DocHeading";
+import DocToc from "./DocToc";
+import RenderRoutes from "./RenderRoutes";
+
+export {
+    ComponentPreview,
+    Attribute,
+    PxScript,
+    DocHeading,
+    DocToc,
+    RenderRoutes
+};