+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added section for colors in utilities.
+- Added progress tracker component and documentation.
 
 ### Changed